authorization added

+++ src/tasks/controller.py
@@ -21,19 +21,22 @@ def create_task(body:TaskSchema, db:Session, user_id:int):
     db.refresh(new_task)
     return new_task
 
-def get_tasks(db:Session):
+def get_tasks(db:Session, user_id:int):
     tasks = db.query(TaskModel).all()
     return tasks
 
 def get_one_task(task_id:int, db:Session) -> HTTPException | type[TaskModel]:
     return get_task_or_404(db, task_id)
 
-def update_task(task_id:int, body:TaskSchema, db:Session)-> type[TaskModel]:
+def update_task(task_id:int, body:TaskSchema, db:Session, user_id:int)-> type[TaskModel]:
     one_task = get_task_or_404(db, task_id)
     # one_task.title = body.title
     # one_task.description = body.description
     # one_task.is_completed = body.is_completed
-
+    print("debug 2")
+    if one_task.user_id != user_id:
+        raise HTTPException(status_code=404, detail="You are not allowed to Update this task")
+    print("debug 3")
     body = body.model_dump()
     for field, value in body.items():
         setattr(one_task, field, value)
@@ -42,8 +45,10 @@ def update_task(task_id:int, body:TaskSchema, db:Session)-> type[TaskModel]:
     db.refresh(one_task)
     return one_task
 
-def delete_task(task_id:int, db:Session):
+def delete_task(task_id:int, db:Session, user_id:int):
     one_task = get_task_or_404(db, task_id)
+    if one_task.user_id != user_id:
+        raise HTTPException(status_code=404, detail="You are not allowed to Delete this task")
     db.delete(one_task)
     db.commit()
     return None
